Duplicate "conforms to the"
Removing duplicate.

However. I seem to remember having seen this error before. The draft seems to have started from an earlier draft of the DI spec, which means that all the possible changes (editorial or otherwise) may have been lost. It would be important to check this section by section...